fix(git): move util function to component
