+++ app/react/portainer/gitops/utils.ts
@@ -32,22 +32,3 @@ export function confirmEnableTLSVerify() {
       'Enabling the verification of TLS certificates without ensuring the correct configuration of your Certificate Authority (CA) for self-signed certificates can result in deployment failures.',
   });
 }
-
-export function cleanGitRepoUrl(url: string) {
-  return url
-    .trim() // remove leading and trailing whitespace
-    .replace(/\/$/, '') // if there's a trailing slash, remove it
-    .replace(/\.git$/, ''); // if there's a trailing .git extension, remove it
-}
-
-export function getGitRepoCommitUrl(url: string, hash: string) {
-  const cleanedUrl = cleanGitRepoUrl(url);
-
-  if (cleanedUrl.startsWith('https://bitbucket.org')) {
-    return `${cleanedUrl}/commits/${hash}`;
-  }
-
-  // this is a fallback for any other git repo
-  // the tested repo includes gitlab, github, and azure devops
-  return `${cleanedUrl}/commit/${hash}`;
-}
